First somewhat ILCDirac inclusion
Added code that (theoretically) allows DESY users to run the simulation and reconstruction on ILCDirac. First somewhat working version, not fully tested, some known bugs to fix.

+++ WWZZSeparation/scripts/run_complete_analysis.sh
@@ -39,6 +39,14 @@ case $i in
 	RUN_PLOTTING=false
     shift # past argument=value
     ;;
+    --UseILCDirac)
+    RUN_ON_ILCDIRAC=true
+    shift
+    ILCSOFT_VERSION=$1
+    shift
+    ILDCONFIG_VERSION=$1
+    shift
+    ;;
 	-h|--help)
 	echo ""
 	echo "In ./run_complete_analysis.sh: executed w/ asking for help:"
@@ -46,8 +54,13 @@ case $i in
 	echo "Script to run complete simulation, reconstruction and analysis of 2nu4q events."
 	echo "Usage: ./run_complete_analysis.sh [-N=number_of_events] [--DetectorModel=detector]"
 	echo "                                  [--no-{ddsim,pandora,analysis,plotting}]"
+	echo "                                  [--UseILCDirac ilcsoft_version ildconfig_version]"
 	echo ""
 	echo "The --no-... flags suppress the ... part of the process."
+  echo ""
+  echo "When using ILCDIRAC the versions of iLCsoft and ILDConfig have to be supplied."
+  echo "They must be versions known to the grid."
+  echo ""
 	echo "Further input information can be set in the input_config.sh script."
 	echo ""
 	echo "Exiting now."
@@ -56,6 +69,7 @@ case $i in
     ;;
     *)
           # unknown option
+    shift
     ;;
 esac
 done
@@ -63,8 +77,8 @@ done
 WORKING_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 STEERING_DIRECTORY=${WORKING_DIR}/../processor/steering_files
-OUTPUT_DIRECTORY=${WORKING_DIR}/../output/${DETECTOR_MODEL}
-# If not already here create detector model specific output directory 
+OUTPUT_DIRECTORY=${WORKING_DIR}/../output
+# If not already here create detector model specific output directory
 if [[ ! -d ${OUTPUT_DIRECTORY} ]] ; then
 	mkdir ${OUTPUT_DIRECTORY}
 fi
@@ -75,13 +89,12 @@ TEMPLATE=${STEERING_DIRECTORY}/template.xml
 
 N_STATES=3
 FINAL_STATES=("vvxyyx" "vvxxxx" "vvyyyy")
-FILES=($(./../input_config.sh filepaths)) 
+FILES=($(./../input_config.sh filepaths))
 CROSS_SECTIONS=($(./../input_config.sh cross-sections))
 
 PIDs=() #Process IDs keep track of my background processes
-
 # If needed and not already here load ILDConfig from GitHub
-if $RUN_DDSIM || $RUN_PANDORA; then
+if ! $RUN_ON_ILCDIRAC && ($RUN_DDSIM || $RUN_PANDORA); then
 	if [[ ! -d ${WORKING_DIR}/ILDConfig ]] ; then
 		echo "Get ILDConfig:"
 		git clone https://github.com/iLCSoft/ILDConfig.git
@@ -89,77 +102,94 @@ if $RUN_DDSIM || $RUN_PANDORA; then
 	fi
 fi
 
-
-
 for (( i=0; i<$(( $N_STATES )); i++ )) do
 	final_state=${FINAL_STATES[$i]}
 	file=${FILES[$i]}
-	cross_section=${CROSS_SECTIONS[$i]}
+  cross_section=${CROSS_SECTIONS[$i]}
 
 	echo "Starting final state: " ${final_state}
-	echo "" 
+	echo ""
 
 	{ # Brackets allow each process to be run in parallel (until plotting they are independent)
 	{
 
-	#--------------------- RUN SIM AND RECO ------------------------#
-	if $RUN_DDSIM && $RUN_PANDORA; then 
-		./run_sim_and_rec.sh -N=${N_EVENTS} --DetectorModel=${DETECTOR_MODEL} --stdhepInput=${file} --processName=${final_state}	
-	elif $RUN_DDSIM; then
-		./run_sim_and_rec.sh -N=${N_EVENTS} --DetectorModel=${DETECTOR_MODEL} --stdhepInput=${file} --processName=${final_state} --no-pandora
-	elif $RUN_PANDORA; then
-		./run_sim_and_rec.sh -N=${N_EVENTS} --DetectorModel=${DETECTOR_MODEL} --stdhepInput=${file} --processName=${final_state} --no-ddsim
-	fi
-		
-	# Clean up unneeded files
-	rm ${OUTPUT_DIRECTORY}/${final_state}*.root 
+  dst_output_file=''
 
-	dst_output_file=${OUTPUT_DIRECTORY}/${final_state}_DST.slcio
+  #--------------------- RUN SIM AND RECO ------------------------#
+  if $RUN_ON_ILCDIRAC; then
+    if $RUN_DDSIM && $RUN_PANDORA; then
+      # Run
+      cd ${WORKING_DIR}/ilcdirac_user_jobs
+      python submit_jobs.py -N ${N_EVENTS} --DetectorModel ${DETECTOR_MODEL} --stdhepInput ${file} --processName ${final_state} --ILCSoftVer ${ILCSOFT_VERSION} --ILDConfigVer ${ILDCONFIG_VERSION}
+      cd ${OUTPUT_DIRECTORY}
+    fi
+    # Retrieve output
+    output_subdir=${ILCSOFT_VERSION}_ILDConfig_${ILDCONFIG_VERSION}_${final_state}
+    if [[ ! -d ${output_subdir} ]] ; then
+      mkdir ${output_subdir}
+    fi
+    cd ${output_subdir}
+    dirac-repo-retrieve-jobs-output-data -r ${WORKING_DIR}/ilcdirac_user_jobs/${ILCSOFT_VERSION}_${ILDCONFIG_VERSION}_${final_state}_${DETECTOR_MODEL}.rep
+    dst_output_file=( $( find $(pwd) -type f -name "*DST.slcio" ) ) # List of absolute paths
+  else
+    if $RUN_DDSIM && $RUN_PANDORA; then
+      ./run_sim_and_rec.sh -N=${N_EVENTS} --DetectorModel=${DETECTOR_MODEL} --stdhepInput=${file} --processName=${final_state}
+    elif $RUN_DDSIM; then
+      ./run_sim_and_rec.sh -N=${N_EVENTS} --DetectorModel=${DETECTOR_MODEL} --stdhepInput=${file} --processName=${final_state} --no-pandora
+    elif $RUN_PANDORA; then
+      ./run_sim_and_rec.sh -N=${N_EVENTS} --DetectorModel=${DETECTOR_MODEL} --stdhepInput=${file} --processName=${final_state} --no-ddsim
+    fi
 
+    # Clean up unneeded files
+    rm ${OUTPUT_DIRECTORY}/${final_state}*.root
+
+    dst_output_file=${OUTPUT_DIRECTORY}/${final_state}_DST.slcio
+  fi
 
 	#--------------------- RUN ANALYSIS ----------------------------#
 	tmp_steering_dir=${STEERING_DIRECTORY}/tmp_${final_state}
 	steering_file=${tmp_steering_dir}/My_${final_state}.xml
 
-	output_root_name=${OUTPUT_DIRECTORY}/${final_state}.root 
+	output_root_name=${OUTPUT_DIRECTORY}/${final_state}.root
 
-	if $RUN_ANALYSIS; then 
+	if $RUN_ANALYSIS; then
 		echo "Starting analysis."
-		echo "" 	
+		echo ""
 
 		if [[ ! -d ${tmp_steering_dir} ]] ; then
 			mkdir ${tmp_steering_dir}
 		fi
 		cp ${TEMPLATE} ${steering_file}
-		
+
 		# Set the input and output file names in the steering file
-		sed -i "116s\.*\ ${cross_section}\ " ${steering_file} 
+    sed -i "116s\.*\ ${cross_section}\ " ${steering_file}
 		sed -i "113s\.*\ ${output_root_name} \  " ${steering_file}
 		sed -i "12s\.*\ ${dst_output_file}\ " ${steering_file}
-		
+
 		cd ${tmp_steering_dir}
-		
+
 		# Start processor that runs analysis of reconstructed objects
 		Marlin My_${final_state}.xml
-		
+
 		cd ${WORKING_DIR}
 		# Clean up unneeded files
 		rm -r ${tmp_steering_dir}
 	fi
 
-	rm ${OUTPUT_DIRECTORY}/${final_state}*.slcio 
+	#rm ${OUTPUT_DIRECTORY}/${final_state}*.slcio
 
 	#Do logging for each final state
-	} > ${LOG_DIRECTORY}/${final_state}.log 2>${LOG_DIRECTORY}/${final_state}.err	
+	} > ${LOG_DIRECTORY}/${final_state}.log 2>${LOG_DIRECTORY}/${final_state}.err
 	} & PIDs+=($!) # Add backround process ID to ID list so I can check if it's still running
 
 done
 
 # Go through list of background process IDs and if one is still running wait
 for pid in "${PIDs[@]}"
-do 
+do
 	wait $pid
 done
+
 
 echo "Simulation, reconstruction and Marlin processors finished."
 echo ""
